platform/x86: ideapad-laptop: move ACPI helpers from header to source file
[ Upstream commit 6c1fa8edfef815a97db57f30216265bfa152792d ]

Since moving ymc_trigger_ec from lenovo-ymc to ideapad-laptop, only the
latter uses these definitions and functions.

Move them back to source file.

Link: https://lore.kernel.org/r/57a48d2582b567f6c6fbcd3b379e17aee0fb5a94.1721898747.git.soyer@irl.hu
Stable-dep-of: 5808c3421695 ("platform/x86: ideapad-laptop: use usleep_range() for EC polling")
(cherry picked from commit 6e7af3d4f67fe8cde5c3a3c6a936af8103304040)
